job_portal/requirements.txt
asgiref==3.8.1
automium==0.2.6
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-environ==0.11.2
fake-useragent==1.5.1
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
legacy==0.1.7
MarkupSafe==2.1.5
oauthlib==3.2.2
psycopg2-binary==2.9.9
requests==2.32.3
requests-oauthlib==2.0.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
useragent==0.1.1
Werkzeug==3.0.3
